swa mixed-chunk retract and glm5.2 tp mtp to extra
