+++ test/registered/sessions/test_streaming_session_swa.py
@@ -1,3 +1,10 @@
+"""Per-commit streaming-session tests on a hybrid-SWA model.
+
+Mirrors the split in test_streaming_session.py: the default config, one retract
+variant, and the abort-leak repro stay per-commit; the mixed-chunk retract
+variant lives in test_streaming_session_extra.py.
+"""
+
 import unittest
 
 from sglang.test.ci.ci_register import register_cuda_ci
@@ -9,20 +16,12 @@ from sglang.test.server_fixtures.streaming_session_fixture import (
     ABORT_REPRO_CHUNKED_PREFILL_SIZE,
     ABORT_REPRO_CONTEXT_LEN,
     ABORT_REPRO_PAGE_SIZE,
+    SWA_COMMON_ARGS,
+    SWA_MODEL,
     StreamingSessionServerBase,
 )
 
-register_cuda_ci(est_time=519, stage="base-b", runner_config="1-gpu-large")
-
-
-SWA_MODEL = "openai/gpt-oss-20b"
-
-# Common gpt-oss-20b launch args. Matches TestSessionLatency/TestSWARadixCacheKL.
-SWA_COMMON_ARGS = [
-    "--mem-fraction-static",
-    "0.70",
-    "--cuda-graph-backend-prefill=disabled",
-]
+register_cuda_ci(est_time=390, stage="base-b", runner_config="1-gpu-large")
 
 
 class TestStreamingSessionSWA(StreamingSessionServerBase, StreamingSessionKitMixin):
@@ -43,21 +42,6 @@ class TestStreamingSessionSWARetractLargePage(
         "4096",
         "--page-size",
         "256",
-        *SWA_COMMON_ARGS,
-    ]
-    env_overrides = [("SGLANG_TEST_RETRACT", True)]
-
-
-class TestStreamingSessionSWARetractMixedChunk(
-    StreamingSessionServerBase, StreamingSessionKitMixin
-):
-    """SWA under retract decode with --enable-mixed-chunk."""
-
-    model = SWA_MODEL
-    extra_args = [
-        "--chunked-prefill-size",
-        "128",
-        "--enable-mixed-chunk",
         *SWA_COMMON_ARGS,
     ]
     env_overrides = [("SGLANG_TEST_RETRACT", True)]
